Centralize Redis connections into a single shared client, reduce MongoDB pool sizes for serverless
Six separate modules created their own Redis connections to Upstash
(middleware.js had two), each maintaining an independent TCP connection.
Under serverless load every concurrent invocation multiplied these
connections, exhausting the database connection budget.

MongoDB pool sizes of 100 (main) and 30 (SSE) were too large for
serverless -- each cold start allocated a pool that never got used,
while concurrent instances collectively exceeded the server's connection
limit. The pools now use 10 and 5 respectively with timeouts and no
minimum idle connections.
