+++ lib/ai/groq.js
@@ -2,16 +2,11 @@ import { z } from "zod";
 import crypto from "crypto";
 import { AppError, ValidationError } from "@/lib/errors";
 import { fetchWithTimeout } from "@/lib/fetch-with-timeout";
-import { Redis } from "@upstash/redis";
+import { getRedis } from "@/lib/redis";
 import { buildSecureMessages } from "@/utils/promptGuard";
 
 function getCacheClient() {
-  const url = process.env.UPSTASH_REDIS_REST_URL;
-  const token = process.env.UPSTASH_REDIS_REST_TOKEN;
-  if (url && token) {
-    return new Redis({ url, token });
-  }
-  return null;
+  return getRedis();
 }
 
 const CACHE_TTL_SECONDS = 3600;
