+++ lib/lockManager.js
@@ -1,21 +1,5 @@
-import { Redis } from "@upstash/redis";
+import { getRedis } from "./redis";
 import logger from "@/utils/logger";
-
-let redisClient;
-
-export function getRedis() {
-  if (
-    !redisClient &&
-    process.env.UPSTASH_REDIS_REST_URL &&
-    process.env.UPSTASH_REDIS_REST_TOKEN
-  ) {
-    redisClient = new Redis({
-      url: process.env.UPSTASH_REDIS_REST_URL,
-      token: process.env.UPSTASH_REDIS_REST_TOKEN,
-    });
-  }
-  return redisClient;
-}
 
 /**
  * Acquire a distributed lock.
