+++ lib/mongodb.js
@@ -1,7 +1,10 @@
 import { MongoClient } from "mongodb";
 
 const options = {
-  maxPoolSize: 100,
+  maxPoolSize: 10,
+  minPoolSize: 0,
+  serverSelectionTimeoutMS: 5000,
+  connectTimeoutMS: 5000,
 };
 
 let client;
@@ -63,7 +66,10 @@ const clientPromise = {
 let sseClient;
 let sseClientPromise;
 const sseOptions = {
-  maxPoolSize: 30,
+  maxPoolSize: 5,
+  minPoolSize: 0,
+  serverSelectionTimeoutMS: 5000,
+  connectTimeoutMS: 5000,
 };
 
 /**
@@ -172,6 +178,23 @@ export async function connectDbForSSE() {
       `Failed to establish database connection: ${error.message}`
     );
   }
+}
+
+export async function disconnectDb() {
+  const clientToClose = client;
+  mongoClientPromise = null;
+  cachedPromise = null;
+  indexesEnsured = false;
+  if (process.env.NODE_ENV === "development") {
+    global._mongoClientPromise = null;
+  }
+  if (clientToClose) {
+    await clientToClose.close().catch(() => {});
+  }
+}
+
+export async function disconnectDbForSSE() {
+  resetSseClient();
 }
 
 export default clientPromise;
