+++ lib/rateLimit.js
@@ -1,23 +1,11 @@
 import { connectDb } from "./mongodb";
-import { Redis } from "@upstash/redis";
+import { getRedis } from "./redis";
 import logger from "@/utils/logger";
 
 const RATE_LIMIT_WINDOW_MS = 60 * 1000;
 const MAX_REQUESTS_PER_WINDOW = 10;
 
 let indexEnsured = false;
-
-let redisClient;
-
-function getRedis() {
-  if (!redisClient) {
-    redisClient = new Redis({
-      url: process.env.UPSTASH_REDIS_REST_URL,
-      token: process.env.UPSTASH_REDIS_REST_TOKEN,
-    });
-  }
-  return redisClient;
-}
 
 async function ensureIndexes(collection) {
   if (indexEnsured) return;
