+++ lib/sessionManager.js
@@ -1,4 +1,4 @@
-import { getRedis } from "./lockManager";
+import { getRedis } from "./redis";
 import { randomUUID } from "crypto";
 
 // Returns true if session management should be bypassed
