+++ package.json
@@ -117,9 +117,8 @@
     "postinstall-postinstall": "^2.1.0",
     "tailwindcss": "^4",
     "tw-animate-css": "^1.3.8",
-    "vitest": "^4.1.8"
     "vite": "6.2.0",
-    "vitest": "3.2.4"
+    "vitest": "^4.1.8"
   },
   "overrides": {
     "serialize-javascript": "7.0.5",
